refactor: 장소 생성 및 조회 기능 위치 변경
