+++ src/main/java/com/culture/ticketing/show/application/dto/PlaceResponse.java
@@ -1,6 +1,6 @@
-package com.culture.ticketing.place.application.dto;
+package com.culture.ticketing.show.application.dto;
 
-import com.culture.ticketing.place.domain.Place;
+import com.culture.ticketing.show.domain.Place;
 import lombok.Getter;
 
 import java.math.BigDecimal;
